refactor state page
