+++ website/frontend/src/components/pages/State/children/StateForm/types.ts
@@ -1,0 +1,5 @@
+import { PayloadValue } from 'components/common/Form/FormPayload/types';
+
+export type FormValues = {
+  payload: PayloadValue;
+};
