+++ website/frontend/src/components/pages/State/types.ts
@@ -1,11 +1,5 @@
 import { Hex } from '@gear-js/api';
 
-import { PayloadValue } from 'components/common/Form/FormPayload/types';
-
-export type FormValues = {
-  payload: PayloadValue;
-};
-
 export type PageParams = {
   programId: Hex;
 };
